+++ CHANGES.rst
@@ -21,6 +21,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- #457: Fix cell export to again use UMTS as the radio type string.
+
 - Optimize various tasks by doing batch queries and inserts.
 
 - Avoid using a metric tag called `name` as it conflicts with a default tag.